moved all component files to their respective folders
