+++ client/src/App.js
@@ -3,10 +3,10 @@ import {useState,useRef} from 'react';
 import { NavbarContainer } from './components/Navbar/Navbar';
 import {Banner} from './components/Banner/Banner';
 import {Skills} from './components/Skills/Skills'
-import {Projects} from './components/Projects';
-import {Contact} from './components/Contact';
-import {Footer} from './components/Footer';
-import {ModalEmail} from './components/Modal';
+import {Projects} from './components/Projects/Projects';
+import {Contact} from './components/Contact/Contact';
+import {Footer} from './components/Footer/Footer';
+import {ModalEmail} from './components/Modal/Modal';
 import 'bootstrap/dist/css/bootstrap.min.css';
