+++ client/src/common/constants/index.js
@@ -9,6 +9,21 @@ import node from '../../assets/img/nodejs.svg';
 import mysql from '../../assets/img/mysql.svg';
 import redux from '../../assets/img/redux.svg';
 
+
+
+import eCommerce from "../../assets/img/ecommerce.png";
+import grandHotel from "../../assets/img/grand_hotel.png";
+import hamburgerMenu from "../../assets/img/hamburger_menu.png";
+
+import invisibleCard from "../../assets/img/invisible_card.png";
+import customSignupForm from "../../assets/img/custom_signup.png";
+import furnitureStore from "../../assets/img/furniture_store.png";
+
+import ticTacToe from "../../assets/img/tic-tac-toe.png";
+
+import profileCard from "../../assets/img/profile_card.png";
+import projImg3 from "../../assets/img/project-img3.png";
+
 export const skills = [
     {skillName:'React',icon:<img alt="react" src={react}/>,rating:4.5,percentage:70},
     {skillName:'Angular',icon:<img alt="angular" src={angular}/>,rating:3,percentage:70},
@@ -23,3 +38,69 @@ export const skills = [
     export const projectDescription = 'These projects where built while I was learning new tech stack and making myself familiar with it. Some project links do actually work so check them out.';
     export const skillDescription = 'Technology stack that I have learnt and used while sailing the path of programming. Feel free to ask me any doubt regarding the below mentioned stack and hopefully I would be able to answer them.';
     export const aboutme = `Graduated with a Master's Degree in Computer Science from the University of Cincinnati. Have 5+ years of professional experience as a Full-stack developer. I love to play around UI and like solving Data structure and algorithm challenges. When I'm not working I will be either Gaming, Traveling, Filming, Gyming,.. you get the idea :)`;
+
+    export const projects = [
+        {
+          title: "E-commerce Website",
+          description: "Stack: React, Redux, JS, Node, Express",
+          isProjectLive:true,
+          projectUrl:'https://crwn-sumoz-live.herokuapp.com/',
+          imgUrl: eCommerce,
+        },
+        {
+          title: "DevConnecter",
+          description: "Stack: React, Redux, JS, Node, Express, MongoDB",
+          isProjectLive:false,
+          projectUrl:'https://calm-lowlands-69902.herokuapp.com/',
+          imgUrl: projImg3,
+        },
+        {
+          title: "Tic-Tac-Toe",
+          description: "Stack: React, JS, CSS",
+          isProjectLive:true,
+          projectUrl:'https://tic-tac-toe-sumoz.herokuapp.com/',
+          imgUrl: ticTacToe,
+        },
+        {
+          title: "Furniture Store",
+          description: "Stack: HTML, CSS, JavaScript",
+          isProjectLive:true,
+          projectUrl:'https://ghewadesumit.github.io/Furniture-Store/#',
+          imgUrl: furnitureStore,
+        },
+        {
+          title: "Custom Sign-up Form",
+          description: "Stack: HTML, CSS, JavaScript",
+          isProjectLive:true,
+          projectUrl:'https://ghewadesumit.github.io/Sumoz-Sign-Up-Form/',
+          imgUrl: customSignupForm,
+        },
+        {
+          title: "Hamburger Menu",
+          description: "Stack: HTML, CSS, JavaScript",
+          isProjectLive:true,
+          projectUrl:'https://ghewadesumit.github.io/sumoz-hamburger-menu/',
+          imgUrl: hamburgerMenu,
+        },
+        {
+          title: "Invisible Card",
+          description: "Stack: HTML, CSS, JavaScript",
+          isProjectLive:true,
+          projectUrl:'https://ghewadesumit.github.io/invisiblecard/',
+          imgUrl: invisibleCard,
+        },
+        {
+          title: "Profile Card",
+          description: "Stack: HTML, CSS, JavaScript",
+          isProjectLive:true,
+          projectUrl:'https://ghewadesumit.github.io/sumoz-profilecard/',
+          imgUrl: profileCard,
+        },
+        {
+          title: "Grand Hotel",
+          description: "Stack: HTML, CSS, JavaScript",
+          isProjectLive:true,
+          projectUrl:'https://ghewadesumit.github.io/flex-grand-hotel/',
+          imgUrl: grandHotel,
+        },
+      ];
