+++ client/src/components/Modal/Modal.scss
@@ -1,0 +1,18 @@
+/*** Modal **/
+.modal-container{
+    position:fixed;
+    width:100vw;
+    height:100vh;
+    display:flex;
+    justify-content: center;
+    align-items: center;
+    background:rgb(200,200,200);
+    z-index: 1000;
+  }
+  
+  .modal-wrapper{
+    width:500px;
+    height:500px;
+    background:white;
+  
+  }
