+++ client/src/components/Navbar/Navbar.js
@@ -3,7 +3,7 @@ import {Navbar, Nav, Container} from 'react-bootstrap';
 import {Download} from 'react-bootstrap-icons';
 import 'animate.css';
 import download from 'downloadjs';
-import {SocialIcons} from '../SocialIcons';
+import {SocialIcons} from '../SocialIcons/SocialIcons';
 import './Navbar.scss';
 
 export const NavbarContainer = ()=>{
