+++ client/src/components/Skills/SkillStarRating.scss
