+++ client/src/components/SocialIcons/SocialIcons.scss
@@ -1,0 +1,3 @@
+.social-icon{
+    margin-top:20px;
+  }
