--- client/src/components/projectData.js
@@ -1,83 +1,0 @@
-import eCommerce from "../assets/img/ecommerce.png";
-import grandHotel from "../assets/img/grand_hotel.png";
-import hamburgerMenu from "../assets/img/hamburger_menu.png";
-
-import invisibleCard from "../assets/img/invisible_card.png";
-import customSignupForm from "../assets/img/custom_signup.png";
-import furnitureStore from "../assets/img/furniture_store.png";
-
-import ticTacToe from "../assets/img/tic-tac-toe.png";
-
-import profileCard from "../assets/img/profile_card.png";
-import projImg3 from "../assets/img/project-img3.png";
-
-
-export const projects = [
-    {
-      title: "E-commerce Website",
-      description: "Stack: React, Redux, JS, Node, Express",
-      isProjectLive:true,
-      projectUrl:'https://crwn-sumoz-live.herokuapp.com/',
-      imgUrl: eCommerce,
-    },
-    {
-      title: "DevConnecter",
-      description: "Stack: React, Redux, JS, Node, Express, MongoDB",
-      isProjectLive:false,
-      projectUrl:'https://calm-lowlands-69902.herokuapp.com/',
-      imgUrl: projImg3,
-    },
-    {
-      title: "Tic-Tac-Toe",
-      description: "Stack: React, JS, CSS",
-      isProjectLive:true,
-      projectUrl:'https://tic-tac-toe-sumoz.herokuapp.com/',
-      imgUrl: ticTacToe,
-    },
-    {
-      title: "Furniture Store",
-      description: "Stack: HTML, CSS, JavaScript",
-      isProjectLive:true,
-      projectUrl:'https://ghewadesumit.github.io/Furniture-Store/#',
-      imgUrl: furnitureStore,
-    },
-    {
-      title: "Custom Sign-up Form",
-      description: "Stack: HTML, CSS, JavaScript",
-      isProjectLive:true,
-      projectUrl:'https://ghewadesumit.github.io/Sumoz-Sign-Up-Form/',
-      imgUrl: customSignupForm,
-    },
-    {
-      title: "Hamburger Menu",
-      description: "Stack: HTML, CSS, JavaScript",
-      isProjectLive:true,
-      projectUrl:'https://ghewadesumit.github.io/sumoz-hamburger-menu/',
-      imgUrl: hamburgerMenu,
-    },
-    {
-      title: "Invisible Card",
-      description: "Stack: HTML, CSS, JavaScript",
-      isProjectLive:true,
-      projectUrl:'https://ghewadesumit.github.io/invisiblecard/',
-      imgUrl: invisibleCard,
-    },
-    {
-      title: "Profile Card",
-      description: "Stack: HTML, CSS, JavaScript",
-      isProjectLive:true,
-      projectUrl:'https://ghewadesumit.github.io/sumoz-profilecard/',
-      imgUrl: profileCard,
-    },
-    {
-      title: "Grand Hotel",
-      description: "Stack: HTML, CSS, JavaScript",
-      isProjectLive:true,
-      projectUrl:'https://ghewadesumit.github.io/flex-grand-hotel/',
-      imgUrl: grandHotel,
-    },
-  ];
-
-
-
-
